Major refactoring and new features
- Refactor philips.py into modular devices/ package (70 classes split into base.py, legacy.py, new_gen.py, new2_gen.py)
- Add network scan feature for automatic device discovery
- Add filter alert binary sensor and event notification
- Add config flow menu for scan/manual device setup
- Reduce cyclomatic complexity (18 → 10)
- Add ruff linter configuration
- Add test suite structure
- Update French and English translations
