+++ .flake8
@@ -1,6 +1,6 @@
 [flake8]
 max-line-length = 100
-max-complexity = 18
+max-complexity = 10
 select = B,C,E,F,W,T4,B9
 per-file-ignores=__init__.py:F401,F403
 # Stuff we ignore thanks to black: https://github.com/ambv/black/issues/429
