+++ custom_components/philips_airpurifier_coap/devices/__init__.py
@@ -1,0 +1,151 @@
+"""Philips AirPurifier device classes."""
+
+from ..const import FanModel
+from .base import (
+    PhilipsEntity,
+    PhilipsGenericControlBase,
+    PhilipsGenericFanBase,
+)
+from .legacy import (
+    PhilipsAC29xx,
+    PhilipsAC303x,
+    PhilipsAC305x,
+    PhilipsAC385x50,
+    PhilipsAC385x51,
+    PhilipsAC1214,
+    PhilipsAC2729,
+    PhilipsAC2889,
+    PhilipsAC2936,
+    PhilipsAC2939,
+    PhilipsAC2958,
+    PhilipsAC2959,
+    PhilipsAC3033,
+    PhilipsAC3036,
+    PhilipsAC3039,
+    PhilipsAC3055,
+    PhilipsAC3059,
+    PhilipsAC3259,
+    PhilipsAC3829,
+    PhilipsAC3836,
+    PhilipsAC4236,
+    PhilipsAC4550,
+    PhilipsAC4558,
+    PhilipsAC5659,
+    PhilipsAC5660,
+    PhilipsAC385450,
+    PhilipsAC385451,
+    PhilipsAC385850,
+    PhilipsAC385851,
+    PhilipsAC385883,
+    PhilipsAC385886,
+    PhilipsGenericFan,
+)
+from .new2_gen import (
+    PhilipsAC085011C,
+    PhilipsAC085020C,
+    PhilipsAC085031C,
+    PhilipsAC085041C,
+    PhilipsAC085070C,
+    PhilipsAC085081,
+    PhilipsAC0950,
+    PhilipsAC0951,
+    PhilipsAC32xx,
+    PhilipsAC3210,
+    PhilipsAC3220,
+    PhilipsAC3221,
+    PhilipsAC3420,
+    PhilipsAC3421,
+    PhilipsAC3737,
+    PhilipsAC4220,
+    PhilipsAC4221,
+    PhilipsAMF765,
+    PhilipsAMF870,
+    PhilipsAMFxxx,
+    PhilipsCX3120,
+    PhilipsCX3550,
+    PhilipsCX5120,
+    PhilipsHU1509,
+    PhilipsHU1510,
+    PhilipsHU5710,
+    PhilipsNew2GenericFan,
+)
+from .new_gen import (
+    PhilipsAC085011,
+    PhilipsAC085020,
+    PhilipsAC085031,
+    PhilipsAC085041,
+    PhilipsAC085070,
+    PhilipsAC085085,
+    PhilipsAC1715,
+    PhilipsNewGenericFan,
+)
+
+model_to_class = {
+    FanModel.AC0850_11: PhilipsAC085011,
+    FanModel.AC0850_11C: PhilipsAC085011C,
+    FanModel.AC0850_20: PhilipsAC085020,
+    FanModel.AC0850_20C: PhilipsAC085020C,
+    FanModel.AC0850_31: PhilipsAC085031,
+    FanModel.AC0850_31C: PhilipsAC085031C,
+    FanModel.AC0850_41: PhilipsAC085041,
+    FanModel.AC0850_41C: PhilipsAC085041C,
+    FanModel.AC0850_70: PhilipsAC085070,
+    FanModel.AC0850_70C: PhilipsAC085070C,
+    FanModel.AC0850_81: PhilipsAC085081,
+    FanModel.AC0850_85: PhilipsAC085085,
+    FanModel.AC0950: PhilipsAC0950,
+    FanModel.AC0951: PhilipsAC0951,
+    FanModel.AC1214: PhilipsAC1214,
+    FanModel.AC1715: PhilipsAC1715,
+    FanModel.AC2729: PhilipsAC2729,
+    FanModel.AC2889: PhilipsAC2889,
+    FanModel.AC2936: PhilipsAC2936,
+    FanModel.AC2939: PhilipsAC2939,
+    FanModel.AC2958: PhilipsAC2958,
+    FanModel.AC2959: PhilipsAC2959,
+    FanModel.AC3033: PhilipsAC3033,
+    FanModel.AC3036: PhilipsAC3036,
+    FanModel.AC3039: PhilipsAC3039,
+    FanModel.AC3055: PhilipsAC3055,
+    FanModel.AC3059: PhilipsAC3059,
+    FanModel.AC3210: PhilipsAC3210,
+    FanModel.AC3220: PhilipsAC3220,
+    FanModel.AC3221: PhilipsAC3221,
+    FanModel.AC3259: PhilipsAC3259,
+    FanModel.AC3420: PhilipsAC3420,
+    FanModel.AC3421: PhilipsAC3421,
+    FanModel.AC3737: PhilipsAC3737,
+    FanModel.AC3829: PhilipsAC3829,
+    FanModel.AC3836: PhilipsAC3836,
+    FanModel.AC3854_50: PhilipsAC385450,
+    FanModel.AC3854_51: PhilipsAC385451,
+    FanModel.AC3858_50: PhilipsAC385850,
+    FanModel.AC3858_51: PhilipsAC385851,
+    FanModel.AC3858_83: PhilipsAC385883,
+    FanModel.AC3858_86: PhilipsAC385886,
+    FanModel.AC4220: PhilipsAC4220,
+    FanModel.AC4221: PhilipsAC4221,
+    FanModel.AC4236: PhilipsAC4236,
+    FanModel.AC4550: PhilipsAC4550,
+    FanModel.AC4558: PhilipsAC4558,
+    FanModel.AC5659: PhilipsAC5659,
+    FanModel.AC5660: PhilipsAC5660,
+    FanModel.AMF765: PhilipsAMF765,
+    FanModel.AMF870: PhilipsAMF870,
+    FanModel.CX3120: PhilipsCX3120,
+    FanModel.CX5120: PhilipsCX5120,
+    FanModel.CX3550: PhilipsCX3550,
+    FanModel.HU1509: PhilipsHU1510,
+    FanModel.HU1510: PhilipsHU1510,
+    FanModel.HU5710: PhilipsHU5710,
+}
+
+__all__ = [
+    "PhilipsEntity",
+    "PhilipsGenericControlBase",
+    "PhilipsGenericFanBase",
+    "PhilipsGenericFan",
+    "PhilipsNewGenericFan",
+    "PhilipsNew2GenericFan",
+    "model_to_class",
+]
