+++ custom_components/philips_airpurifier_coap/fan.py
@@ -11,7 +11,7 @@ from homeassistant.helpers.entity import Entity
 
 from .config_entry_data import ConfigEntryData
 from .const import DOMAIN
-from .philips import model_to_class
+from .devices import model_to_class
 
 _LOGGER = logging.getLogger(__name__)
