+++ custom_components/philips_airpurifier_coap/humidifier.py
@@ -18,7 +18,7 @@ from homeassistant.helpers.entity import Entity
 
 from .config_entry_data import ConfigEntryData
 from .const import DOMAIN, HUMIDIFIER_TYPES, FanAttributes, FanFunction
-from .philips import PhilipsGenericControlBase, model_to_class
+from .devices import PhilipsGenericControlBase, model_to_class
 
 _LOGGER = logging.getLogger(__name__)
